Splited headers to the separate classes for code readability.
